+++ src/Components/Projects/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "./Projects.module.css";
+import styles from "./index.module.css";
 import { Link } from "react-router-dom";
 
 const Projects = () => {
