+++ src/Components/about/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "./About.module.css";
+import styles from "./index.module.css";
 
 const About = () => {
   return (
