+++ src/Components/contactUs/index.module.css
@@ -1,6 +1,4 @@
 .container{
-    /* width: 1190px;
-    height: 603px; */
     margin-left: 150px;
 }
