+++ src/Components/footer/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "./Footer.module.css";
+import styles from "./index.module.css";
 
 const Footer = () => {
   return (
