+++ src/Components/home/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "./Home.module.css";
+import styles from "./index.module.css";
 
 const Home = () => {
   return (
