+++ src/Components/home/index.module.css
@@ -74,41 +74,3 @@
     margin-top: 50px;
 }
 
-/* @media screen and (max-width : 830px) {
-    .container {
-        display: block;
-        margin-top: auto;
-    }
-
-    .title,
-    .text {
-        margin-left: 0px;
-    }
-
-    .image {
-        display: flex;
-        height: 53px;
-        width: 52.95px;
-        margin-left: 0px;
-    }
-
-    .number {
-        margin-left: 50px;
-    }
-
-    .right img {
-        height: 470px;
-        width: 470px;
-        z-index: 1;
-    }
-
-    .button {
-        padding: 0px;
-        margin-left: 500px;
-        width: 200px;
-        z-index: 1;
-        position: absolute;
-        right: 584px;
-        top: 655px;
-    }
-} */
