+++ src/Components/navBar/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "./NavBar.module.css";
+import styles from "./index.module.css";
 import { Outlet,Link } from "react-router-dom";
 // import { BrowserRouter as Router, Route } from "react-router-dom";
