+++ src/Components/navBar/index.module.css
@@ -1,7 +1,6 @@
 .container {
     font-family: Verdana, Geneva, Tahoma, sans-serif;
     font-variant: small-caps;
-    /* padding: 30px; */
     height: 102px;
     width: 1440px;
 }
@@ -35,19 +34,3 @@ li :hover{
     border-bottom: 1px solid black;
 }
 
-/* @media screen and (max-width:830px) {
-    img {
-        height: 43.7px;
-        width: 70px;
-        margin-left: 15px;
-        margin-top: auto;
-    }
-    .items {
-        float: none;
-        display: flex;
-        gap: 10px;
-        flex-wrap: wrap;
-        width: 100%;
-    }
-
-} */
